auth

+++ backend/controllers/productController.js
@@ -1,11 +1,14 @@
 import {sql} from "../config/db.js";
 
 
+
+
 export const getAllProducts = async (req,res)=>{
     try {
-        const products = await sql `SELECT * FROM products ORDER BY created_at DESC`;
+        const userId = req.user.id;
+        const products = await sql `SELECT * FROM products WHERE user_id=${userId} ORDER BY created_at DESC`;
         console.log("Fetched Products",products);
-        res.status(200).json({success:true,data:products});
+        res.status(200).json({success:true,data:products,name:req.username});
         
     } catch (error) {
         console.log("Error getProducts",error);
@@ -16,8 +19,9 @@ export const getAllProducts = async (req,res)=>{
 
 export const getProduct = async(req,res)=>{
     const {id} = req.params;
+    const userId = req.user.id;
     try {   
-        const product  = await sql `SELECT * FROM products WHERE id=${id}`;
+        const product  = await sql `SELECT * FROM products WHERE user_id=${userId} AND id=${id}`;
         if(product.length == 0){
             return res.status(404).json({
                 success:false,
@@ -37,12 +41,13 @@ export const getProduct = async(req,res)=>{
 
 export const createProduct = async (req,res) =>{
     const {name,price,image} = req.body;
+    const userId = req.user.id;
     
     if(!name || !price || !image){
         return res.status(400).json({success:false,message:"All fields are required"});
     }
     try { 
-        const newProduct = await sql `INSERT INTO products (name,price,image) VALUES (${name},${price},${image}) RETURNING *`;
+        const newProduct = await sql `INSERT INTO products (name,price,image,user_id) VALUES (${name},${price},${image},${userId}) RETURNING *`;
         console.log("new product added", newProduct);
         res.status(201).json({success:true,data:newProduct[0]});
     } catch (error) {
@@ -56,8 +61,9 @@ export const createProduct = async (req,res) =>{
 export const updateProduct = async (req,res)=>{
     const {id} = req.params;
     const {name,price,image} = req.body;
+    const userId = req.user.id;
     try {
-        const updatedProduct = await sql `UPDATE products SET name=${name},price=${price},image=${image} WHERE id=${id} RETURNING *`;
+        const updatedProduct = await sql `UPDATE products SET name=${name},price=${price},image=${image} WHERE user_id=${userId} AND id=${id} RETURNING *`;
         if(updatedProduct.length == 0){
             return res.status(404).json({
                 success:false,
@@ -75,8 +81,9 @@ export const updateProduct = async (req,res)=>{
 
 export const deleteProduct = async (req,res)=>{
     const {id} = req.params;
+    const userId = req.user.id;
     try {
-        const deletedProduct = await sql `DELETE FROM products where id=${id} RETURNING *`;
+        const deletedProduct = await sql `DELETE FROM products where user_id=${userId} AND id=${id} RETURNING *`;
 
         if(deletedProduct.length==0){
             return res.status(404).json({
